+++ details.html
@@ -29,20 +29,20 @@
   </header>
   <main class="d-flex justify-content-center align-items-center">
     <div class="card mb-3 cardinfo my-5">
-      <div class="row g-0">
+      <div class="row g-0" aria-hidden="true">
         <div class="col-md-4">
-          <img src="images/recursos_Amazing_Events_Task_1/cinema.jpg" class="img-fluid rounded-start h-100" alt="...">
+          <img src=".." class="img-fluid rounded-start h-100 card-img-top" alt="...">
         </div>
         <div class="col-md-8">
           <div class="card-body">
-            <h5 class="card-title">Avengers</h5>
-            <p class="card-text">Marvel's Avengers 3d premiere the start of an epic saga with your best superheroes.</p>
-            <p class="card-category">Category: Let's go to the cinema</p>
-            <p class="card-place">Place: Room D1</p>
-            <p class="card-capacity">Capacity: 9000</p>
-            <p class="card-estimate">Estimate: 9000</p>
-            <p class="card-price">Price: 250</p>
-            <p class="card-date"><small class="text-body-secondary">Date: 2023-10-15</small></p>
+            <h5 class="card-title"></h5>
+            <p class="card-text"></p>
+            <p class="card-category"></p>
+            <p class="card-place"></p>
+            <p class="card-capacity"></p>
+            <p class="card-estimate"></p>
+            <p class="card-price"></p>
+            <p class="card-date"><small class="text-body-secondary"></small></p>
           </div>
         </div>
       </div>
