+++ home.html
@@ -136,8 +136,7 @@
 
     </footer>
 
-    <script src="module.js" ></script>
-    <script src="home.js"></script>
+    <script type="module" src="./home.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
     crossorigin="anonymous"></script>
