+++ pastEvents.html
@@ -101,7 +101,7 @@
     
             <div class="searchvar">
                 <form class="d-flex justify-content-end" role="search">
-                    <input class="form-control me-2 w-50" type="search" placeholder="Search" aria-label="Search" id="searchbarhome"> 
+                    <input class="form-control me-2 w-50" type="search" placeholder="Search" aria-label="Search" id="searchbarpast"> 
                     <button class="btn btn-outline-dark" type="submit">Search</button>
                   </form>
     
@@ -136,7 +136,7 @@
     </footer>
 
 
-    <script src="pastEvents.js"></script>
+    <script type="module" src="./pastEvents.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
     crossorigin="anonymous"></script>
